# src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: Courses API
  description: API for course management
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /courses:
    post:
      summary: Create course
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCoursePayload'
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoursePayload'

    get:
      summary: Get all courses
      operationId: getCourses
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetCoursePayload'

  /courses/{id}:
    get:
      summary: Get course by id
      operationId: getCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoursePayload'

    put:
      summary: Update course
      operationId: updateCourse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoursePayload'
      responses:
        '200':
          description: Updated course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoursePayload'

components:
  schemas:
    CreateCoursePayload:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string

    UpdateCoursePayload:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    LessonPayload:
      type: object
      properties:
#        TODO for now dummy object, please specify what exactly it should include
        id:
          type: integer
          format: int64
        title:
          type: string

    GetCoursePayload:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/LessonPayload'